remove WriteObserver interface and rename AndroidWriteObserver to take its place
