+++ android/src/main/java/com/msopentech/thali/universal/toronionproxy/OnionProxyContext.kt
@@ -13,7 +13,6 @@ See the Apache 2 License for the specific language governing permissions and lim
 package com.msopentech.thali.universal.toronionproxy
 
 import android.os.Process
-import com.msopentech.thali.android.toronionproxy.AndroidWriteObserver
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.BufferedWriter
@@ -165,5 +164,5 @@ class OnionProxyContext(
 
     @Throws(IOException::class)
     fun generateWriteObserver(file: File): WriteObserver =
-        AndroidWriteObserver(file)
+        WriteObserver(file)
 }
