+++ android/src/main/java/com/msopentech/thali/universal/toronionproxy/WriteObserver.kt
@@ -1,4 +1,19 @@
 /*
+Copyright (C) 2011-2014 Sublime Software Ltd
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+ */
+/*
 Copyright (c) Microsoft Open Technologies, Inc.
 All Rights Reserved
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
@@ -12,17 +27,32 @@ See the Apache 2 License for the specific language governing permissions and lim
 */
 package com.msopentech.thali.universal.toronionproxy
 
+import android.os.FileObserver
+import java.io.File
+import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 
 /**
- * Android uses FileObserver and Java uses the WatchService, this class abstracts the two.
+ * Adapted from the Briar WriteObserver code
  */
-interface WriteObserver {
-    /**
-     * Waits timeout of unit to see if file is modified
-     * @param timeout How long to wait before returning
-     * @param unit Unit to wait in
-     * @return True if file was modified, false if it was not
-     */
-    fun poll(timeout: Long, unit: TimeUnit): Boolean
+class WriteObserver(file: File) : FileObserver(file.absolutePath, CLOSE_WRITE) {
+
+    private val countDownLatch = CountDownLatch(1)
+
+    fun poll(timeout: Long, unit: TimeUnit): Boolean =
+        try {
+            countDownLatch.await(timeout, unit)
+        } catch (e: InterruptedException) {
+            throw RuntimeException("Internal error has caused AndroidWriteObserver to not be reliable.", e)
+        }
+
+    override fun onEvent(i: Int, s: String) {
+        stopWatching()
+        countDownLatch.countDown()
+    }
+
+    init {
+        require(file.exists()) { "FileObserver doesn't work properly on files that don't already exist." }
+        startWatching()
+    }
 }
